requirements.txt
streamlit
langchain
langchain-community
python-dotenv
sentence-transformers
langchain-huggingface
langchain_groq
groq
langchain-text-splitters
pypdf
faiss-cpu
langgraph
easyocr
pillow
opencv-python-headless

